ECS grid methods and map partials
Atomic PRs moment.
